refactor(formatter): split formatter.rs into emit, render, summary, pagination modules (#1221)
formatter.rs (2,886L) is replaced by a formatter/ module directory with four
focused sub-modules:

- emit.rs      -- low-level output builders (strip_base_path, format_file_info_parts,
                  format_file_entry, aggregate_dir_stats, render_top_files_section,
                  def_use_write_read_counts, render_def_use_group, snippet_one_line)
- render.rs    -- per-entity rendering (format_file_details, format_file_details_summary,
                  format_module_info, format_classes_section, format_imports_section,
                  FormatterError)
- summary.rs   -- directory and focused symbol aggregates (format_structure, format_summary,
                  format_focused_internal, format_focused_summary_internal,
                  format_focused_summary)
- pagination.rs -- cursor and pagination helpers (format_structure_paginated,
                   format_file_details_paginated, format_focused_paginated,
                   format_chains_as_tree)

mod.rs re-exports all public symbols so all six production import sites
(analyze.rs, formatter_defuse.rs, tools/analyze_{file,directory,symbol,module}.rs)
require no changes -- their existing aptu_coder_core::formatter::* imports continue
to resolve through the facade.

idempotency tests updated: hardcoded src/formatter.rs path changed to
src/formatter/mod.rs now that the file is a module directory.

Closes #1221 (F2)
